refactor: Settings 화면 조립을 route 경계로 옮긴다
